requirements.txt
absl-py==1.2.0
asgiref==3.7.2
asttokens==2.0.8
backcall==0.2.0
bitstring==3.1.9
branca==0.6.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
Django==4.2.2
django-betterforms==2.0.0
django-bootstrap-modal-forms==3.0.3
django-cleanup==7.0.0
django-static-jquery==2.1.4
ecdsa==0.18.0
esptool==4.5.1
executing==1.1.1
ExifRead==3.0.0
ezdxf==1.0.3
filelock==3.9.0
folium==0.14.0
fonttools==4.37.4
geographiclib==2.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
gpsphoto==2.2.3
grpcio==1.49.1
idna==3.4
imutils==0.5.4
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.2.0
platformdirs==2.6.2
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
reedsolo==1.6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.5.1
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.17.1
wcwidth==0.2.5
Werkzeug==2.2.2
